功能新增

+++ app01/models.py
@@ -5,7 +5,7 @@ from django.db import models
 
 class UserInfo(models.Model):
     name = models.CharField(verbose_name='姓名', max_length=10)
-    phone = models.IntegerField(verbose_name='电话')
+    phone = models.BigIntegerField(verbose_name='电话')
 
     def __unicode__(self):
         return self.name
